Corrigido tipo de dado. Gerado nova Migration

+++ Entra21.MiAuDota/Entra21.MiAuDota.Repositorio/Entidades/Animal.cs
@@ -11,7 +11,7 @@ namespace Entra21.MiAuDota.Repositorio.Entidades
         public string Alergias { get; set; }
         public string? OutrasInformacoesMedicas { get; set; }
         public string Foto { get; set; }
-        public int Idade { get; set; }
+        public byte Idade { get; set; }
         public double Peso { get; set; }
         public double Altura { get; set; }
         public bool Castrado { get; set; }
